refactor: CommunityDetail 페이지에서 불필요한 로직 제거 및 FeedContent 컴포넌트로 분리
