+++ apps/client/src/widgets/community/components/feed-content/feed-content.css.ts
