--- apps/client/src/widgets/community/constant/detail-content.ts
@@ -1,1 +1,0 @@
-export const POINT = '•';
